requirements.txt
pytest
selene==2.0.b8
allure-pytest